refactor(metrics): keep the Python transport out of the stubs
The generated endstone/metrics/__init__.pyi covered the Python Metrics,
MetricsBase and MetricsConfig as well as the bound charts. griffe copies
annotation text verbatim, so every annotation those classes carried had to be
written as typing.Any or abc.ABC to reach the stub with an import behind it,
and a bare Plugin needed a rewrite rule of its own.

None of that is worth the churn it puts through files the charts changed for
other reasons. The stub now describes the bindings, as every other one does,
and base.py and config.py go back to the annotations they had. Both are
renamed with a leading underscore, since nothing outside the package should
reach past endstone.metrics for them.
